Change: directory structure has changed, layout components move to layout directory
