+++ src/components/HeroSlider/HeroSlider.tsx
@@ -1,5 +1,3 @@
-import { useState } from 'react'
-
 export const HeroSlider = () => {
   return <div>HeroSlider</div>
 }
